requirements.txt
gevent
requests
python-nmap
dnspython
pyfiglet